novo layout da tela de exploração espacial e modificação na tela de menu dos planetas
