+++ app/src/main/java/com/renan/digitalspace/favorite/view/FavoriteFragment.kt
@@ -106,7 +106,7 @@ class FavoriteFragment : Fragment(), IFavorite {
         val btnBackView = _view.findViewById<ImageButton>(R.id.ibBackFavorite)
 
         btnBackView.setOnClickListener {
-            _navController.navigate(R.id.action_favoriteFragment_to_exploracaoFragment)
+            _navController.navigate(R.id.action_favoriteFragment_to_explorationFragment)
         }
     }
