+++ app/src/main/res/layout/fragment_favorite.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/bg_planetas"
+    android:background="@drawable/bg_planetas2"
     android:orientation="vertical"
     tools:context=".favorite.view.FavoriteFragment">
