+++ app/src/main/res/layout/fragment_quiz_questions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    android:background="@drawable/bg_planetas"
+    android:background="@drawable/bg_planetas2"
     tools:context=".quiz.view.QuizQuestionsFragment">
