+++ app/src/main/res/layout/item_planets_menu.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal" android:layout_width="wrap_content"
+    android:orientation="horizontal"
+    android:layout_width="wrap_content"
     android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/imgPlanetRecycler"
         android:layout_width="wrap_content"
-        android:layout_height="@dimen/dim130dp"
+        android:layout_height="@dimen/dim120dp"
         android:adjustViewBounds="true"
         android:src="@drawable/sol"/>
