+++ app/src/main/res/values/strings.xml
@@ -71,8 +71,14 @@
     <string name="updateMessage"> Última atualização de imagem: </string>
     <string name="esqueci_minha_senha">Esqueci minha senha</string>
     <string name="bem_vindos_exploradores">Bem vindos exploradores! Vistam seus trajes e preparem-se para a decolagem. Iniciaremos uma missão de exploração pelo Sistema Solar. Acomodem-se em seus assentos, apertem os cintos e preparem-se para ir, audaciosamente, onde ninguém jamais esteve. Partimos em 3, 2, 1...</string>
+    <string name="exploration_text">Bem vindo, explorador!\nVista seu traje, aperte o cinto e prepare-se para a decolagem.
+        Iniciaremos uma missão de exploração pelo Sistema Solar.\nPartimos em 3, 2, 1...</string>
     <string name="quiz">Quiz</string>
     <string name="jogar">Jogar</string>
     <string name="pr_xima">Próxima</string>
     <string name="teste_seus_conhecimentos">Teste aqui seus conhecimentos sobre o Sistema Solar!</string>
+    <string name="milkyway_imgdescription">Foto da via láctea</string>
+    <string name="solareclipse_imgdescription">Foto de um eclipse solar</string>
+    <string name="earth_imgdescription">Foto do planeta Terra</string>
+
 </resources>
